Recurrent Payments Base

+++ src/main/java/com/mozido/recurrentpayments/entity/ScheduledRecurrentPayment.java
@@ -28,7 +28,7 @@ public class ScheduledRecurrentPayment {
     private String username;
     private String companyCode;
 
-    private BigDecimal amount;
+    private Double amount;
     private LocalDate startDate;
     private LocalDate endDate;
     private int endAfter;
@@ -78,11 +78,11 @@ public class ScheduledRecurrentPayment {
         this.userId = userId;
     }
 
-    public BigDecimal getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
